Proyecto final para instalar en IDE local

+++ src/main/java/curso/spring/controller/CategoriasController.java
@@ -15,50 +15,63 @@ import curso.spring.model.Categorias;
 import curso.spring.model.Productos;
 import curso.spring.model.Usuarios;
 import curso.spring.service.CategoriasService;
+import curso.spring.service.ProductosService;
 
+/**
+ * CRUD para entidad Categorias
+ * @author Gonzalo
+ *
+ */
 @Controller
-@RequestMapping("/categorias")
+@RequestMapping("/category")
 public class CategoriasController {
 	
 	@Autowired
 	CategoriasService catS;
 	
+	@Autowired
+	ProductosService pS;
+	
 
 	@GetMapping("/list")
 	public String listCategorias(Model model) {
 		model.addAttribute("list", catS.getAllCategorias());
-		return "categorias/list";
+		return "category/list";
 	}
 	
 	@GetMapping("/add")
 	public String addcat(Model model) {
 		model.addAttribute("categoria", new Categorias());
-		return "categorias/new";
+		return "category/new";
 	}
 	
 	@PostMapping("/add/submit")
 	public String createCat(@ModelAttribute Categorias categoria) {
 		catS.addCategoria(categoria);
-		return "redirect:/categorias/list";
+		return "redirect:/category/list";
 	}
 	
 	@GetMapping("/delete/{id}")
 	public String deleteCat(@PathVariable int id) {
-		catS.deleteCategoria(id);
-		return "redirect:/categorias/list";
+		
+		if (pS.getProductsByCat(id).isEmpty()) {
+			catS.deleteCategoria(id);
+		}
+		
+		return "redirect:/category/list";
 	}
 	
 	@GetMapping("/edit/{id}")
 	public String actualizacionForm(@PathVariable int id, Model model) {
 		Categorias c = catS.getCategoriaById(id);
 		model.addAttribute("categoria", c);
-		return "categorias/edit";
+		return "category/edit";
 	}
 	
 	@PostMapping("/edit/submit")
 	public String actualizar(@ModelAttribute Categorias categoria) {
 		catS.editCategoria(categoria);
-		return "redirect:/categorias/list";
+		return "redirect:/category/list";
 
 	}
